% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{ClarabelSolver}
\alias{ClarabelSolver}
\title{A persistent Clarabel solver that can be reused across R calls.}
\format{
An object of class \code{clarabel::ClarabelSolver__bundle} (inherits from \code{savvy_clarabel__sealed}) of length 1.
}
\usage{
ClarabelSolver
}
\description{
A persistent Clarabel solver that can be reused across R calls.
}
\keyword{datasets}
