% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarabel.R
\name{solver_status_descriptions}
\alias{solver_status_descriptions}
\title{Return the solver status description as a named character vector}
\usage{
solver_status_descriptions()
}
\value{
a named list of solver status descriptions, in order of status codes returned by the solver
}
\description{
Return the solver status description as a named character vector
}
\examples{
solver_status_descriptions()[2] ## for solved problem
solver_status_descriptions()[8] ## for max iterations limit reached
}
